# Makefile: disable swagger server