--- a/.github/workflows/check.yaml
+++ b/.github/workflows/check.yaml
@@ -25,6 +25,6 @@ jobs:
             ${{ runner.os }}-golang
       - name: Make Check
         run: |
+          make check
           make build
           make tools
-          make check