--- a/tests/client/Makefile
+++ b/tests/client/Makefile
@@ -23,6 +23,7 @@ static: install-tools
 
 tidy:
 	@ go mod tidy
+	git diff go.mod go.sum | cat
 	git diff --quiet go.mod go.sum
 
 test: enable-codegen dashboard-ui