… go.sum
… tests/client/go.sum